Chore: capsulate redis operations into respository module
