+++ src/models.rs
@@ -1,54 +1,6 @@
+use crate::repository::{DisplayOptions, Repository, Segment, Timer};
 use serde::{Deserialize, Serialize};
-use crate::color::Color;
 use std::sync::Arc;
-use tokio::sync::broadcast;
-
-//main.rs
-fn default_zero() -> u32 {
-    0
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct Segment {
-    label: String,
-    time: u32,
-    sound: bool,
-    color: Option<Color>,
-    #[serde(default = "default_zero")]
-    count_to: u32,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub enum PreStartBehaviour {
-    ShowZero,
-    RunNormally,
-}
-
-impl Default for PreStartBehaviour {
-    fn default() -> Self {
-        PreStartBehaviour::ShowZero
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Default, Debug)]
-pub struct DisplayOptions {
-    #[serde(default)]
-    clock: bool,
-    #[serde(default)]
-    pre_start_behaviour: PreStartBehaviour,
-}
-
-#[derive(Serialize, Deserialize, Clone, Default, Debug)]
-pub struct Timer {
-    // Return after TimerRequest
-    pub segments: Vec<Segment>,
-    pub repeat: bool,
-    pub display_options: Option<DisplayOptions>,
-    pub start_at: u64,
-    pub stop_at: Option<u64>,
-    pub password: String,
-    pub id: String, // 5 random chars
-}
 
 #[derive(Serialize, Clone)]
 #[serde(tag = "type", content = "data")]
@@ -60,16 +12,15 @@ pub enum DonationMethod {
 pub struct InstanceProperties {
     pub demo: bool,
     pub donation: Option<Vec<DonationMethod>>,
+    pub s3_host: String,
 }
 
 pub type SharedState = Arc<AppState>;
 pub struct AppState {
-    pub redis: redis::aio::ConnectionManager,
+    pub repository: Repository,
     pub jwt_key: String,
     pub instance_properties: InstanceProperties,
-    pub redis_task_rx: broadcast::Receiver<Timer>,
 }
-
 
 //timer.rs
